+++ src/main/java/ureca/nolmung/jpa/diaryplace/DiaryPlace.java
@@ -1,24 +1,24 @@
-package ureca.nolmung.jpa.dogDiary;
+package ureca.nolmung.jpa.diaryplace;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import ureca.nolmung.jpa.config.BaseEntity;
 import ureca.nolmung.jpa.diary.Diary;
-import ureca.nolmung.jpa.dog.Dog;
-import ureca.nolmung.jpa.user.User;
+import ureca.nolmung.jpa.place.Place;
 
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-public class DogDiary {
+public class DiaryPlace extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "dog_diary_id")
+    @Column(name = "diary_place_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
@@ -26,6 +26,6 @@ public class DogDiary {
     private Diary diary;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "dog_id")
-    private Dog dog;
+    @JoinColumn(name = "place_id")
+    private Place place;
 }
